Segmentsystem mal nochmal komplett ueberarbeitet libsgl gesaeubert

+++ helper/sgldisplist.h
@@ -46,20 +46,20 @@ public:
 	list<shared_obj>	Objects;
 	GLuint		*Objects_CW;
 	GLuint		*Objects_CCW;
-	unsigned int	ObjCnt_CW,ObjCnt_CCW;
-	unsigned int	ObjSize_CW,ObjSize_CCW;
+	unsigned short	ObjCnt_CW,ObjCnt_CCW;
+	unsigned short	ObjSize_CW,ObjSize_CCW;
 	void	CallAllLists();
 	void	Compile(bool force=false);
 	bool	check_recompile,check_sorting;
 	bool	renderTransparent,render_non_native;
 	void	ListInfo();
-	unsigned int isThere(SGLObjBase *obj);
+	unsigned short isThere(SGLObjBase *obj);
 
 private:
 	bool	removeOb_CW(GLuint ListID);
 	bool	removeOb_CCW(GLuint ListID);
 	template<class T>
-	bool	grow(T *&liste,unsigned int &size,unsigned int newsize);
+	bool	grow(T *&liste,unsigned short &size,unsigned short newsize);
 };
 
 #endif
